# src/backend/docs/v1/v1_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# _В основе лежит мой курсовой проект 😎_",
        "title": "Идентификация на КПП",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/infocard-photos/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Метод для получения элемента коллекции фотографий сотрудников",
                "produces": [
                    "image/jpeg",
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получение элемента коллекции фотографий сотрудников",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор информационной карточки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фотография сотрудника успешно получена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Авторизация неуспешна",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    },
                    "404": {
                        "description": "Карточка не найдена",
                        "schema": {
                            "$ref": "#/definitions/http.StatusNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка получения фотографии пользователя",
                        "schema": {
                            "$ref": "#/definitions/http.StatusInternalServerError"
                        }
                    }
                }
            }
        },
        "/infocards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Метод для получения коллекции информационных карточек",
                "tags": [
                    "admin"
                ],
                "summary": "Получение коллекции информационных карточек",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Значение для фильтрации",
                        "name": "pattern",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поле для фильтрации и сортировки",
                        "name": "field",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Направление сортировки",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о карточках успешно получена",
                        "schema": {
                            "$ref": "#/definitions/admin.listFullInfoCardsResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Авторизация неуспешна",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка получения карточек пользователей",
                        "schema": {
                            "$ref": "#/definitions/http.StatusInternalServerError"
                        }
                    }
                }
            }
        },
        "/infocards/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Метод для получения элемента коллекции информационных карточек",
                "tags": [
                    "admin"
                ],
                "summary": "Получение элемента коллекции информационных карточек",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор информационной карточки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о карточке успешно получена",
                        "schema": {
                            "$ref": "#/definitions/admin.getFullInfoCardResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Авторизация неуспешна",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    },
                    "404": {
                        "description": "Карточка не найдена",
                        "schema": {
                            "$ref": "#/definitions/http.StatusNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка получения карточки пользователя",
                        "schema": {
                            "$ref": "#/definitions/http.StatusInternalServerError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Метод для подтверждения информационной карточки сотрудника",
                "tags": [
                    "admin"
                ],
                "summary": "Подтверждение информационной карточки сотрудника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор информационной карточки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о карточке успешно подтверждена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/http.StatusBadRequest"
                        }
                    },
                    "401": {
                        "description": "Авторизация неуспешна",
                        "schema": {
                            "$ref": "#/definitions/http.StatusUnauthorized"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.getFullInfoCardResponse": {
            "type": "object",
            "properties": {
                "document": {
                    "$ref": "#/definitions/model.FullDocument"
                },
                "passages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ShortPassage"
                    }
                }
            }
        },
        "admin.listFullInfoCardsResponse": {
            "type": "object",
            "properties": {
                "infoCards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FullInfoCard"
                    }
                }
            }
        },
...
        "http.StatusBadRequest": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Incorrect request body date of birth"
                }
            }
        },
        "http.StatusInternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to register employee"
                }
            }
        },
        "http.StatusNotFound": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Failed to get info card document"
                }
            }
        },
        "http.StatusUnauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Cant login employee"
                }
            }
        },
        "model.DataDocument": {
            "type": "object",
            "properties": {
                "documentType": {
                    "type": "string"
                },
                "serialNumber": {
                    "type": "string"
                }
            }
        },
        "model.FullDocument": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.DataDocument"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.KeyValueField"
                    }
                }
            }
        },
        "model.FullInfoCard": {
            "type": "object",
            "properties": {
                "companyID": {
                    "type": "integer"
                },
                "createdDate": {
                    "type": "string"
                },
                "createdEmployeeID": {
                    "type": "integer"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isConfirmed": {
                    "type": "boolean"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "post": {
                    "type": "string"
                }
            }
        },
        "model.KeyValueField": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.ShortPassage": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}